Did most of practice exercises
