--- practice/src/main/java/ca/jrvs/practice/dataStructure/HelloWorld.java
@@ -1,12 +1,0 @@
-package ca.jrvs.practice.dataStructure;
-
-class HelloWorld {
-
-  // Your program begins with a call to main().
-  // Prints "Hello, World" to the terminal window.
-  public static void main(String args[]) {
-    System.out.println("Hello, World");
-
-
-  }
-}
