+++ practice/src/main/java/ca/jrvs/practice/dataStructure/LambdaStream.java
@@ -1,0 +1,2 @@
+package ca.jrvs.practice.dataStructure;public class LambdaStream {
+}
