+++ practice/src/main/java/ca/jrvs/practice/dataStructure/LambdaStreamExc.java
@@ -1,0 +1,2 @@
+package ca.jrvs.practice.dataStructure;public interface LambdaStreamExc {
+}
