+++ practice/src/main/java/ca/jrvs/practice/dataStructure/RegexExc.java
@@ -1,0 +1,26 @@
+package ca.jrvs.practice.apps;
+
+public interface RegexExc {
+  /**
+   * return true if filename extension is jpg or jpeg (case insensitive)
+   * @param filename
+   * @return
+   */
+  public boolean matchJpeg(String filename);
+
+  /**
+   * return true if ip is valid
+   * to simply the problem, IP address range is from 0.0.0.0 to 999.999.999.999
+   * @param ip
+   * @return
+   */
+  public boolean matchIp(String ip);
+
+  /**
+   * return true if line empty (e.g. empty, white space, tabs, etc.)
+   * @param line
+   * @return
+   */
+  public boolean isEmptyLine(String line);
+
+}
